refactor #7 : wiki, wikipage 패키지 병합 및 클래스 명칭 일부 수정 기존의 wiki, wikipage로 패키지가 나뉘어있던 부분을 도메인 단위인 wiki로 병합하여 관리하도록 수정. WikiPageAndRevisionCommandService를 WikiDomainCommandService로 명칭 변경. WikiBookmarkRepository를 WikiBookmarkCommandRepository로 명칭 변경.
