+++ lpick-backend/src/main/java/com/notfound/lpickbackend/AUTO_ENTITIES/Artist.java
@@ -1,6 +1,6 @@
 package com.notfound.lpickbackend.AUTO_ENTITIES;
 
-import com.notfound.lpickbackend.wikipage.command.application.domain.WikiPage;
+import com.notfound.lpickbackend.wiki.command.application.domain.WikiPage;
 import jakarta.persistence.*;
 import lombok.*;
