+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wiki/command/application/domain/WikiBookmark.java
@@ -1,7 +1,6 @@
-package com.notfound.lpickbackend.wikipage.command.application.domain;
+package com.notfound.lpickbackend.wiki.command.application.domain;
 
 import com.notfound.lpickbackend.AUTO_ENTITIES.UserInfo;
-import com.notfound.lpickbackend.wikipage.command.application.domain.WikiPage;
 import jakarta.persistence.*;
 import lombok.*;
