+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wiki/command/application/domain/WikiStatus.java
@@ -1,0 +1,5 @@
+package com.notfound.lpickbackend.wiki.command.application.domain;
+
+public enum WikiStatus {
+    OPEN, BLIND, DELETE
+}
