+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wiki/query/repository/WikiPageQueryRepository.java
@@ -1,6 +1,6 @@
-package com.notfound.lpickbackend.wikipage.query.repository;
+package com.notfound.lpickbackend.wiki.query.repository;
 
-import com.notfound.lpickbackend.wikipage.command.application.domain.WikiPage;
+import com.notfound.lpickbackend.wiki.command.application.domain.WikiPage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
