+++ lpick-backend/src/main/java/com/notfound/lpickbackend/wiki/query/service/WikiPageQueryService.java
@@ -1,12 +1,12 @@
-package com.notfound.lpickbackend.wikipage.query.service;
+package com.notfound.lpickbackend.wiki.query.service;
 
 import com.notfound.lpickbackend.common.exception.CustomException;
 import com.notfound.lpickbackend.common.exception.ErrorCode;
 import com.notfound.lpickbackend.wiki.command.application.domain.PageRevision;
 import com.notfound.lpickbackend.wiki.query.service.PageRevisionQueryService;
-import com.notfound.lpickbackend.wikipage.command.application.domain.WikiPage;
-import com.notfound.lpickbackend.wikipage.query.dto.WikiPageViewResponse;
-import com.notfound.lpickbackend.wikipage.query.repository.WikiPageQueryRepository;
+import com.notfound.lpickbackend.wiki.command.application.domain.WikiPage;
+import com.notfound.lpickbackend.wiki.query.dto.response.WikiPageViewResponse;
+import com.notfound.lpickbackend.wiki.query.repository.WikiPageQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
