--- lpick-backend/src/main/java/com/notfound/lpickbackend/wikipage/command/application/domain/WikiStatus.java
@@ -1,5 +1,0 @@
-package com.notfound.lpickbackend.wikipage.command.application.domain;
-
-public enum WikiStatus {
-    OPEN, BLIND, DELETE
-}
